Created TaskPolicy, authorized users, updated Task Controller with some logic and DocBlocks

+++ app/Http/Controllers/TaskController.php
@@ -3,28 +3,37 @@
 namespace App\Http\Controllers;
 
 use App\Models\Task;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Contracts\View\View;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Response;
 use Dompdf\Dompdf;
-use Illuminate\Support\Facades\View;
 
 
 class TaskController extends Controller
 {
     /**
-     * Display a listing of the resource.
+     * Display a listing of the resource for user.
+     *
+     * @return View
      */
-    public function index()
+    public function index(): View
     {
-        $tasks = Task::where('user_id', auth()->id())->orderBy('created_at', 'desc')->get();
-
+        if (!is_superadmin(user())){
+            $tasks = Task::where('user_id', auth()->id())->orderBy('created_at', 'desc')->get();
+        }else{
+            $tasks = Task::all();
+        }
         return view('tasks.index', compact('tasks'));
     }
 
     /**
      * Show the form for creating a new resource.
+     *
+     * @return View
      */
-    public function create()
+    public function create(): View
     {
         return view('tasks.create');
 
@@ -32,8 +41,11 @@ class TaskController extends Controller
 
     /**
      * Store a newly created resource in storage.
+     *
+     * @param Request $request
+     * @return RedirectResponse
      */
-    public function store(Request $request)
+    public function store(Request $request): RedirectResponse
     {
         $validatedData = $request->validate([
             'title' => 'required',
@@ -55,26 +67,39 @@ class TaskController extends Controller
 
     /**
      * Display the specified resource.
+     *
+     * @param Task $task
+     * @return View
      */
-    public function show(Task $task)
+    public function show(Task $task): View
     {
-        //
+        $this->authorize('view', $task);
+        return view('tasks.show', compact('task'));
     }
 
     /**
      * Show the form for editing the specified resource.
+     *
+     * @param Task $task
+     * @return View
      */
-    public function edit(Task $task)
+    public function edit(Task $task): View
     {
+        $this->authorize('update', $task);
         return view('tasks.edit', compact('task'));
 
     }
 
     /**
      * Update the specified resource in storage.
+     *
+     * @param Request $request
+     * @param Task $task
+     * @return RedirectResponse
      */
-    public function update(Request $request, Task $task)
+    public function update(Request $request, Task $task): RedirectResponse
     {
+        $this->authorize('update', $task);
         $validatedData = $request->validate([
             'title' => 'required',
             'comment' => 'nullable',
@@ -92,16 +117,25 @@ class TaskController extends Controller
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Remove the specified resource from storage
+     * @param Task $task
+     * @return RedirectResponse
+     * @throws AuthorizationException
      */
-    public function destroy(Task $task)
+    public function destroy(Task $task):RedirectResponse
     {
+        $this->authorize('delete', $task);
         $task->delete();
 
         return redirect()->route('tasks.index')->with('success', 'Task deleted successfully.');
     }
 
-    public function generateReport(Request $request)
+    /**
+     * Generate a report based on the specified date range
+     * @param Request $request
+     * @return \Illuminate\Http\Response
+     */
+    public function generateReport(Request $request): \Illuminate\Http\Response
     {
         // Retrieve tasks based on the date range
         $startDate = $request->input('start_date');
